requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
supabase
python-jose
pinecone
open-deep-research
tavily-python
langchain
langchain-openai
langchain-community
networkx